pyproject.toml
[project]
name = "mobilenet-test"
version = "0.1.0"
description = ""
authors = [
    {name = "genii01",email = "genii0627@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
